requirements.txt
django 
djangorestframework 
sentence-transformers 
faiss-cpu 
requests 
beautifulsoup4
